+++ R/ggpiestats.R
@@ -9,8 +9,6 @@
 #'   of fit test) will be run for the `x` variable. Otherwise an appropriate
 #'   association test will be run. This argument can not be `NULL` for
 #'   `ggbarstats` function.
-#' @param factor.levels A character vector with labels for factor levels of
-#'   `main` variable.
 #' @param title The text for the plot title.
 #' @param caption The text for the plot caption.
 #' @param palette If a character string (e.g., `"Set1"`), will use that named
@@ -25,9 +23,6 @@
 #'   (default), `"counts"`, `"both"`.
 #' @param label.args Additional aesthetic arguments that will be passed to
 #'   `geom_label`.
-#' @param bf.message Logical that decides whether to display a caption with
-#'   results from Bayes Factor test in favor of the null hypothesis (default:
-#'   `FALSE`).
 #' @inheritParams tidyBF::bf_contingency_tab
 #' @inheritParams statsExpressions::expr_contingency_tab
 #' @inheritParams theme_ggstatsplot
@@ -67,20 +62,14 @@
 #' set.seed(123)
 #'
 #' # one sample goodness of fit proportion test
-#' ggstatsplot::ggpiestats(
-#'   data = ggplot2::msleep,
-#'   x = vore,
-#'   perc.k = 1
-#' )
+#' ggstatsplot::ggpiestats(ggplot2::msleep, vore)
 #'
 #' # association test (or contingency table analysis)
 #' ggstatsplot::ggpiestats(
 #'   data = mtcars,
 #'   x = vs,
 #'   y = cyl,
-#'   bf.message = TRUE,
 #'   nboot = 10,
-#'   factor.levels = c("0 = V-shaped", "1 = straight"),
 #'   legend.title = "Engine"
 #' )
 #' @export
@@ -93,7 +82,6 @@ ggpiestats <- function(data,
                        ratio = NULL,
                        paired = FALSE,
                        results.subtitle = TRUE,
-                       factor.levels = NULL,
                        label = "percentage",
                        perc.k = 0,
                        label.args = list(alpha = 1, fill = "white"),
@@ -245,13 +233,6 @@ ggpiestats <- function(data,
 
   # =================================== plot =================================
 
-  # getting labels for all levels of the 'x' variable factor
-  if (is.null(factor.levels)) {
-    legend.labels <- as.character(df %>% dplyr::pull({{ x }}))
-  } else {
-    legend.labels <- factor.levels
-  }
-
   # if no. of factor levels is greater than the default palette color count
   palette_message(
     package = package,
@@ -289,8 +270,7 @@ ggpiestats <- function(data,
     ggplot2::scale_y_continuous(breaks = NULL) +
     paletteer::scale_fill_paletteer_d(
       palette = paste0(package, "::", palette),
-      name = "",
-      labels = unique(legend.labels)
+      name = ""
     ) +
     theme_pie(
       ggtheme = ggtheme,